Search shipments on Newgistics  (#3)
With this commit we'll allow developers to search shipments on Newgistics using the `Newgistics::Shipment class`:
```
shipments = Newgistics::Shipment.where(start_received_timestamp: start_date).
  where(end_received_timestamp: end_date).all
start_received_timestamp and end_received_timestamp are Date types in ISO 8601 format
```
`Newgistics::Shipment.where(conditions).all` will return a list of `Newgistics::Shipment` elements if the request is successful. Otherwise it will raise a QueryError if there is an error or failure in the response.