Allow users to get product inventory from Newgistics (#4)
With this PR we'll allow developers to get inventory information 
from Newgistics using the Newgistics::Product class:

```
products = Newgistics::Product.all
products = Newgistics::Product.where(sku: sku).
  where(warehouse: warehouse_id).all
```

`sku` is a product's `sku` and `warehouse` is the warehouse's id 
number, where both of these values are strings

`Newgistics::Product.all` and `Newgistics::Product.where(conditions).all` 
will return a list of Newgistics::Product elements if the request is 
successful. Otherwise it will raise a `Newgistics::QueryError` if there 
is an error or failure in the response.