Allow users to query inventory details using Newgistics::Inventory (#7)
With this change we'll allow developers to retrieve inventory
details from newgistics using the Newgistics::Inventory class:

````ruby
Newgistics::Inventory.
  where(start_timestamp: start_date, end_timestamp: end_date).
  all
```

`start_date` and `end_date` are Date types in ISO8601 format.

You can use the `where` method to specify the parameters for the
Search. Parameter keys will be automatically camelized when sent to
Newgistics, for a full list of the available parameters refer to
the Newgistics API documentation.