Update cassettes to have human-readable responses
With this change we'll update all of our cassettes so we can
actually read the body of the response, it was previously
being recorded as a binary string which makes it harder for
us to understand the expected outcome of our integration tests.

We'll also start automatically filtering our API key to ensure
we don't commit sensitive information to the repository.