Add time zone to newgistics timestamps
Newgistics API doesn't return a timezone with their timestamps
so we'll allow our gem users to set the timezone that should
be used when parsing timestamps through the configuration.

This will only work as long as the value returned from the API
is a ISO 8601 timestamp, if it isn't then the attribute will
remain as a String.