Update shipment contents (#16)
Update Shipment Contents

Allow developers to update a shipments contents through newgistics
API. You can create a new `Newgistics::Order` with either/both
`add_items` or `remove_items` which are arrays of items to update
the shipments contents with. You can do then call
`order.update` on the order to send the changes to newgistics. If
it was successful the call will return true and the success value
on the order itself will be set to true.