Allow developers to query manifests
With this change we'll allow developers to query manifests
using the `Newgistics::Manifest.where` method.

We'll also refactor our models so they include the
`Newgistics::Model` module, which allows each class to override
the `.element_selector` method to indicate the css selector
we should use to get to the XML node that contains the attributes
for instances of this class when parsing responses from the API.

This simplifies building Newgistics::Query objects and gives us
a bit more flexibility when parsing deeply-nested objects.