Reduce the number of objects allocated for request processing
With this change we'll update our StringHelper class to perform
in-place modifications to strings so we reduce the number of
strings allocated when these methods get called.

Sadly, readability will be affected by this change but I believe
the positive impact this will have on memory usage is worth it,
especially for large requests.