Avoid overriding address fields when updating a shipment's address
With this change we'll prevent developers from accidentally overriding
unwanted fields on a shipment when performing an address update.

Basically, if we only set the :status attribute in a
Newgistics::ShipmentAddressUpdate, then that should be the only
field that gets updated in Newgistics, previously we were
including nil attributes in the request and Newgistics was setting
those fields to nil, which is not what we want.